Refactoring + Bad Smell fix
